[refactor] Create a file for set comparison in assertion
